cover: move media placeholder out of cover
