+++ extensions/editor.js
@@ -6,7 +6,7 @@ import './shared/block-category';
 import './shared/plan-upgrade-notification';
 import './shared/stripe-connection-notification';
 import './shared/external-media';
-import './shared/blocks/cover';
+import './shared/blocks';
 import './shared/premium-blocks';
 import analytics from '../_inc/client/lib/analytics';
